requirements.txt
langchain
langchain-community
langchain-openai
pdfplumber
faiss-cpu
sentence-transformers
python-dotenv
openai